TWINS-302: refactoring code

+++ core/src/main/java/org/twins/core/service/permission/PermissionGrantUserGroupService.java
@@ -15,7 +15,10 @@ import org.springframework.transaction.annotation.Transactional;
 import org.twins.core.dao.domain.DomainEntity;
 import org.twins.core.dao.permission.PermissionGrantUserGroupEntity;
 import org.twins.core.dao.permission.PermissionGrantUserGroupRepository;
+import org.twins.core.exception.ErrorCodeTwins;
 import org.twins.core.service.auth.AuthService;
+import org.twins.core.service.user.UserGroupService;
+import org.twins.core.service.user.UserService;
 
 import java.sql.Timestamp;
 import java.time.Instant;
@@ -31,6 +34,9 @@ public class PermissionGrantUserGroupService extends EntitySecureFindServiceImpl
     private final PermissionGrantUserGroupRepository repository;
     private final AuthService authService;
     private final PermissionSchemaService permissionSchemaService;
+    private final PermissionService permissionService;
+    private final UserGroupService userGroupService;
+    private final UserService userService;
 
     @Override
     public CrudRepository<PermissionGrantUserGroupEntity, UUID> entityRepository() {
@@ -54,9 +60,18 @@ public class PermissionGrantUserGroupService extends EntitySecureFindServiceImpl
 
     @Override
     public boolean validateEntity(PermissionGrantUserGroupEntity entity, EntitySmartService.EntityValidateMode entityValidateMode) throws ServiceException {
-        if (entity.getPermissionSchema() == null)
-            entity.setPermissionSchema(permissionSchemaService.findEntitySafe(entity.getPermissionSchemaId()));
-        return !isEntityReadDenied(entity, EntitySmartService.ReadPermissionCheckMode.none);
+        switch (entityValidateMode) {
+            case beforeSave:
+                if (entity.getPermissionSchema() == null || !entity.getPermissionSchema().getId().equals(entity.getPermissionSchemaId()))
+                    entity.setPermissionSchema(permissionSchemaService.findEntitySafe(entity.getPermissionSchemaId()));
+                if (entity.getPermission() == null || !entity.getPermission().getId().equals(entity.getPermissionId()))
+                    entity.setPermission(permissionService.findEntitySafe(entity.getPermissionId()));
+                if (entity.getUserGroup() == null || !entity.getUserGroup().getId().equals(entity.getUserGroupId()))
+                    entity.setUserGroup(userGroupService.findEntitySafe(entity.getUserGroupId()));
+                if (entity.getGrantedByUserId() != null && (entity.getGrantedByUser() == null || !entity.getGrantedByUser().getId().equals(entity.getGrantedByUserId())))
+                    entity.setGrantedByUser(userService.findEntitySafe(entity.getGrantedByUserId()));
+        }
+        return true;
     }
 
     @Transactional(rollbackFor = Throwable.class)
@@ -82,7 +97,10 @@ public class PermissionGrantUserGroupService extends EntitySecureFindServiceImpl
 
     @Transactional
     public void deleteById(UUID id) throws ServiceException {
-        //todo need validate???
+        PermissionGrantUserGroupEntity entity = findEntitySafe(id);
+        DomainEntity domain = authService.getApiUser().getDomain();
+        if (!entity.getPermissionSchema().getDomainId().equals(domain.getId()))
+            throw new ServiceException(ErrorCodeTwins.DOMAIN_PERMISSION_DENIED);
         entitySmartService.deleteAndLog(id, repository);
     }
 }
